refactor(memory): split schema.rs (1458 lines) into schema/ directory
Extracts the flat schema.rs into three focused sub-modules under
src/openhuman/memory/schema/:

  definitions.rs  — the `schemas()` match (all ControllerSchema literals)
  handlers.rs     — all `handle_*` bridge functions + shared helpers
  registry.rs     — `all_controller_schemas` / `all_registered_controllers`

schema/mod.rs re-exports the original public surface unchanged so all
callers in core::all and memory::mod.rs continue to compile without
modification. schema_tests.rs remains in place, referenced via
#[path = "../schema_tests.rs"].
